+++ plugins/nf-schema/src/main/nextflow/validation/CustomEvaluators/FormatPathEvaluator.groovy
@@ -26,7 +26,7 @@ class FormatPathEvaluator implements Evaluator {
         def String value = node.asString()
 
         // Skip validation of S3 paths for now
-        if (value.startsWith('s3://')) {
+        if (value.startsWith('s3://') || value.startsWith('az://') || value.startsWith('gs://')) {
             log.debug("S3 paths are not supported by 'FormatPathEvaluator': '${value}'")
             return Evaluator.Result.success()
         }